Moving files to differen directory
