+++ src/components/ElementsInfo/Element.js
