+++ src/components/ElementsInfo/Element.module.scss
