+++ src/components/ElementsInfo/Elements.module.scss
